Added operator comments and moved around some functions
