+++ blockpost-frontend/package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "bls-signatures": "^2.0.3",
     "ethers": "^6.13.1",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
